Experiment adding SIMD vector types
The 'MultiN a' type specifies how the array data should be laid out, namely that for a single Multi2 value the two elements of type 'a' should be adjacent in memory.

  t1 :: Multi2 Float
  t1 = [Multi2 f1 f2, Multi2 f3 f4 ...]
    => [f1,f2,f3,f4, ...]

Note that we haven't restricted 'a' to be of primitive types, which is what we actually need to use actual SIMD registers, but this still works out...

  t2 :: Multi2 (Float,Int)
  t3 = [Multi2 (f1,i1) (f2,i2), ...]
    => [f1,f2,...]
       [i1,i2,...]

This worked out for a small test but perhaps actual code generation will be tricky.