Add 'Int' encoding directly into Idx constructor
But hide that fact using pattern synonyms. Adding the Int field was originally proposed by @mwu-tow or one of his team I think, who found it to improve performance in some parts of the compiler, but I can't find the original discussion that describes what they found.

A quick test with lulesh shows that compilations is a bit quicker, while other phases slow down a lot. Perhaps the use of pattern synonyms are not as well optimised yet as regular pattern matches?